Change skater auto force sensor braking to only last a couple seconds

+++ controller/Core/Inc/controller_wireless.h
@@ -18,7 +18,7 @@ typedef struct {
 	UART_HandleTypeDef *uart;
 	uint8_t uart_buffer[10];
 	int message_contents;
-	uint16_t ms_since_comms;
+	uint32_t ms_since_comms;
 } Wireless;
 
 /** PUBLIC FUNCTIONS **/
